+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-ettersending
 
+## 3.33.10
+
+### Patch Changes
+
+-   Updated dependencies [a10b53c]
+    -   @navikt/sif-common-amplitude@2.25.8
+    -   @navikt/sif-common-formik-ds@1.9.4
+    -   @navikt/sif-common-soknad-ds@10.15.12
+    -   @navikt/sif-common-core-ds@9.17.13
+    -   @navikt/sif-common-sentry@0.26.10
+    -   @navikt/sif-common-utils@3.46.6
+
 ## 3.33.9
 
 ### Patch Changes